requirements.txt
Flask>=3.0.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-core>=0.3.70
langchain-google-genai>=0.0.6
langchain-community>=0.0.10
langchain-chroma>=0.2.5
chromadb>=0.4.22
google-generativeai>=0.3.2
